sync commit

+++ dot_product.m
@@ -1,3 +1,3 @@
 function out = dot_product(a,b)
-    out = sum(a.*b)
+    out = int32(sum(a.*b))
 end
